requirements.txt
 # Core dependencies
torch==2.0.1
numpy==1.24.3
pandas==2.0.3

# Clustering
faiss-cpu==1.7.4  # Use faiss-gpu for GPU support

# Configuration
pyyaml==6.0

# Utilities
tqdm==4.65.0
tensorboard==2.13.0
scikit-learn==1.3.0


